Add back base delete functionality
